python/agents/blog-writer/pyproject.toml
[project]
name = "blogger-agent"
version = "0.1.0"
description = "AI-driven agent designed to facilitate the exploration of the blogger agent landscape"
authors = [{ name = "Pier Paolo Ippolito", email = "pierippolito@google.com" }]
license = "Apache-2.0"
readme = "README.md"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-adk>=1.0.0",
    "asyncpg>=0.29.0",
    "psycopg>=3.2.10",
]
requires-python = ">=3.10,<3.13"

[dependency-groups]
dev = [
    "pytest~=8.3.2",
    "pytest-asyncio~=0.23.7",
    "google-adk[eval]>=1.0.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.14.1",
]
deployment = [
    "absl-py~=2.2.1",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.0.20240914",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "W", # pycodestyle warnings
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["blogger_agent"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

[tool.agent-starter-pack]
example_question = "Write an article about Nano Banana"

[tool.agent-starter-pack.settings]
agent_directory = "blogger_agent"
